Move search classes into single datasources file
